change test_app_with_invalid_environment_vars to integration test
